:construction: Improve API
